REFACTORING TO SOC
