+++ modules/globals.py
@@ -1,0 +1,4 @@
+# globals.py
+
+session_store = {}
+current_tasks = {}
